+++ HISTORY.md
@@ -1,5 +1,5 @@
-unreleased
-==========
+1.3.7 / 2022-04-28
+==================
 
   * Fix hanging on large stack of sync routes